+++ src/app/organizer/ItemTable.tsx
@@ -114,7 +114,7 @@ export default function ItemTable({ categories }: { categories: ItemCategoryTree
   const destinyVersion = useSelector(destinyVersionSelector);
   const dispatch = useThunkDispatch();
 
-  const { customTotalStatsByClass } = itemCreationContext;
+  const { customStats } = itemCreationContext;
 
   const classCategoryHash =
     categories.map((n) => n.itemCategoryHash).find((hash) => hash in categoryToClass) ?? 999;
@@ -162,8 +162,6 @@ export default function ItemTable({ categories }: { categories: ItemCategoryTree
     [terminal, items]
   );
 
-  const customStatTotal = customTotalStatsByClass[classIfAny] ?? emptyArray();
-
   const columns: ColumnDefinition[] = useMemo(
     () =>
       getColumns(
@@ -173,7 +171,7 @@ export default function ItemTable({ categories }: { categories: ItemCategoryTree
         itemInfos,
         wishList,
         hasWishList,
-        customStatTotal,
+        customStats,
         loadoutsByItem,
         newItems,
         destinyVersion,
@@ -185,7 +183,7 @@ export default function ItemTable({ categories }: { categories: ItemCategoryTree
       statHashes,
       itemType,
       itemInfos,
-      customStatTotal,
+      customStats,
       classIfAny,
       loadoutsByItem,
       newItems,
@@ -504,9 +502,14 @@ export default function ItemTable({ categories }: { categories: ItemCategoryTree
         return (
           <div
             key={column.id}
-            className={clsx(styles[column.id], styles.header, {
-              [styles.stats]: isStatsColumn,
-            })}
+            className={clsx(
+              styles[column.id],
+              column.id.startsWith('customstat_') && styles.customstat,
+              styles.header,
+              {
+                [styles.stats]: isStatsColumn,
+              }
+            )}
             role="columnheader"
             aria-sort="none"
           >
@@ -645,6 +648,7 @@ function TableRow({
           onClick={onRowClick(row, column)}
           className={clsx(styles[column.id], {
             [styles.hasFilter]: column.filter,
+            [styles.customstat]: column.id.startsWith('customstat_'),
           })}
           role="cell"
         >