+++ src/app/organizer/table-types.ts
@@ -59,3 +59,5 @@ export interface ColumnSort {
   columnId: string;
   sort: SortDirection;
 }
+
+export type ColumnWithStat = ColumnDefinition & { statHash: number };